requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.4.0
apache-airflow==2.9.3
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-postgres==5.11.2
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apache-libcloud==3.8.0
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.14.3
bitsandbytes==0.48.2
blinker==1.8.2
cachelib==0.9.0
cachetools==6.2.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chonkie==1.4.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docling==2.64.0
docling-core==2.54.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
docutils==0.16
duckdb==1.4.2
elastic-transport==8.17.1
elasticsearch==8.19.2
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.13.1
Faker==38.2.0
fastapi==0.124.0
fastapi-mcp==0.4.0
fastuuid==0.14.0
filelock==3.19.1
filetype==1.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flatbuffers==25.9.23
frozenlist==1.4.1
fsspec==2023.12.2
gevent==25.9.1
ggml_py==0.9.4.post1
gliner==0.2.21
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
grand-cypher==1.0.1
grand-graph==0.7.0
grandiso==2.2.0
greenlet==3.3.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
hf-xet==1.2.0
hnswlib==0.8.0
httpcore==1.0.5
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.0
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.7
ImageHash==4.3.2
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflect==7.5.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.2
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lark==1.3.1
latex2mathml==3.78.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
litellm==1.80.9
llama_cpp_python==0.3.16
lockfile==0.12.2
lxml==6.0.2
Mako==1.3.5
markdown-it-py==3.0.0
marko==2.2.1
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mcp==1.23.1
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
ml_dtypes==0.5.4
model2vec==0.7.0
more-itertools==10.3.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.0.5
multiprocess==0.70.18
networkx==3.5
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.20.0
onnxmltools==1.14.0
onnxruntime==1.23.2
openai==2.9.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pandas==2.3.3
pathspec==0.12.1
peft==0.18.0
pendulum==3.0.0
pgvector==0.4.2
pillow==11.3.0
pluggy==1.5.0
polyfactory==3.1.0
prettytable==3.17.0
prison==0.2.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.18.0
PyJWT==2.8.0
pylatexenc==2.10
pyparsing==3.2.5
pypdfium2==4.30.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-nvd3==0.16.0
python-pptx==1.0.2
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.9.0
PyYAML==6.0.1
rapidocr==3.4.2
referencing==0.35.1
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
setproctitle==1.3.3
setuptools==66.1.1
shapely==2.1.2
shellingham==1.5.4
six==1.16.0
skl2onnx==1.19.1
skops==0.13.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlite-vec==0.1.6
sqlparse==0.5.0
sse-starlette==3.0.3
starlette==0.50.0
staticvectors==0.2.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
tika==3.1.0
tiktoken==0.12.0
time-machine==2.14.2
timm==1.0.22
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
ttstokenizer==1.1.0
-e git+https://github.com/neuml/txtai.git@1d7b6a98c905d498d08e85d834e1e043169c1a4f#egg=txtai&subdirectory=../../txtai
typeguard==4.4.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.2.0
urllib3==2.0.7
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webrtcvad-wheels==2.0.14
websocket==0.2.1
websockets==15.0.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
xlsxwriter==3.2.9
xmltodict==1.0.2
yarl==1.22.0
zipp==3.19.2
zope.event==6.1
zope.interface==8.1.1
